#0 hotfix: 카톡 미리보기 에러
