--- src/app/meme-test/memeMetadata.ts
@@ -1,21 +1,0 @@
-import { Metadata } from "next";
-
-export function generateMemeMetadat0a(): Metadata {
-  return {
-    title: "MZ 성향 테스트하기!",
-    openGraph: {
-      title: "MZ 성향 테스트하기!",
-      description: "성향 테스트 결과를 확인해보세요!",
-      url: "https://moo-mool.vercel.app/meme-test",
-      images: [
-        {
-          url: "https://moo-mool.vercel.app/assets/icons/meme-test-home.png",
-          width: 800,
-          height: 600,
-          alt: "밈 테스트 미리보기",
-        },
-      ],
-      type: "website",
-    },
-  };
-}
